+++ src/modes/loopTypewriter.ts
@@ -1,24 +1,16 @@
-import {
-	writeEffect,
-	unwriteEffect,
-	getRandomElement,
-	createElement,
-	cleanChildNodes
-} from '@svelte-typewriter/helpers'
+import { writeEffect, unwriteEffect, getRandomElement } from '@svelte-typewriter/helpers'
 import type { LoopAnimation, TypewriterModeFn } from '@svelte-typewriter/types'
 
 const loopTypewriter: LoopAnimation = async (parentElement, { currentNode, text }, options) => {
-	cleanChildNodes(parentElement)
-	const loopParagraphTag = currentNode.tagName.toLowerCase()
-	const loopParagraph = createElement(text, loopParagraphTag)
-	const loopParagraphAttributes = [...currentNode.attributes]
-	loopParagraphAttributes.forEach(({ name, value }) => loopParagraph.setAttribute(name, value))
-	parentElement.appendChild(loopParagraph)
-	await writeEffect({ currentNode: loopParagraph, text }, options)
+	//const loopParagraphTag = currentNode.tagName.toLowerCase()
+	//const loopParagraph = createElement(text, loopParagraphTag)
+	//const loopParagraphAttributes = [...currentNode.attributes]
+	//loopParagraphAttributes.forEach(({ name, value }) => loopParagraph.setAttribute(name, value))
+	//parentElement.appendChild(loopParagraph)
+	await writeEffect({ currentNode, text }, options)
 	const textWithUnescapedAmpersands = text.replaceAll('&', '&amp;')
-	const fullyWritten = loopParagraph.innerHTML === textWithUnescapedAmpersands
-	fullyWritten && (await unwriteEffect(loopParagraph, options))
-	cleanChildNodes(parentElement)
+	const fullyWritten = currentNode.innerHTML === textWithUnescapedAmpersands
+	fullyWritten && (await unwriteEffect(currentNode, options))
 }
 
 const mode: TypewriterModeFn = async ({ node, elements }, options) => {