+++ CHANGELOG.md
@@ -14,11 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Commits
 
 - build: replace tsc with rollup for building typescript [`96d620d`](https://github.com/henriquehbr/svelte-typewriter/commit/96d620d8ea4a843230fd29b543bc95c232b25c3d)
+- chore(example): add a component for each animation mode [`df06ae4`](https://github.com/henriquehbr/svelte-typewriter/commit/df06ae4fd7cc7a06214bfeec76c178cbb9a7b992)
 - refactor: convert helpers and modes to typescript (#33) [`4683a02`](https://github.com/henriquehbr/svelte-typewriter/commit/4683a02978a2b2608268cf5d136071d1993ad2e4)
 - refactor: add proper types for helpers and modes (#33) [`23ac8bc`](https://github.com/henriquehbr/svelte-typewriter/commit/23ac8bc7800166b7cdd4912ad3a092cde6b3ae46)
 - refactor: convert "utils" into separated "helpers" (#33) [`fb6158d`](https://github.com/henriquehbr/svelte-typewriter/commit/fb6158d296e2f99dde8ec7e5a39a6edf39087911)
 - refactor: unify loop and loopRandom modes [`914ef72`](https://github.com/henriquehbr/svelte-typewriter/commit/914ef722e84152f581546cd7e7bba8f168ad144e)
-- refactor: modularize getRandomElement from loop mode (#33) [`8ba2e5f`](https://github.com/henriquehbr/svelte-typewriter/commit/8ba2e5fa77aff40d34884bf0930a6c2f2133ff1c)
+- refactor: modularize getRandomElement from loop mode (#33) [`6fd55b7`](https://github.com/henriquehbr/svelte-typewriter/commit/6fd55b7c739073bbc49f482073fffbab19a1aed0)
 - refactor(utils): integrate loop logic into "writeEffect" (#33) [`1ba9750`](https://github.com/henriquehbr/svelte-typewriter/commit/1ba975028c1e1b068c231fb03967796f7ad0821e)
 - chore: remove unnecessary "any" types [`0e205bd`](https://github.com/henriquehbr/svelte-typewriter/commit/0e205bd2e1fc4b1cca60dfd6f7ada0a1b1cde490)
 - refactor(helpers): add writeEffect [`939a86b`](https://github.com/henriquehbr/svelte-typewriter/commit/939a86b73780d4b2857fdabb1a7f0f2ed49a4416)